requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
stripe
pytest
httpx
python-multipart
